refactor: introduced key-matchers
