+++ src/lib/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './deep-merge';
 export * from './filtered-data';
 export * from './is-mergeable-object';
-export * from './lenient-matching';
+export * from './key-matchers';
